src/test/java/edu/hw3/Task2Test.java
package edu.hw3;

import java.util.List;
import java.util.stream.Stream;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;
import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

class Task2Test {
    static Stream<Arguments> ClusterizeArguments() {
        return Stream.of(
            Arguments.of("()()()", List.of("()", "()", "()")),
            Arguments.of("((()))", List.of("((()))")),
            Arguments.of("((()))(())()()(()())", List.of("((()))", "(())", "()", "()", "(()())")),
            Arguments.of("((())())(()(()()))", List.of("((())())", "(()(()()))"))
        );
    }

    @ParameterizedTest
    @MethodSource("ClusterizeArguments")
    @DisplayName("Test Clusterize Algorithm")
    void testAtbashEncodeDecode(String input, List<String> output) {
        assertThat(Task2.clusterize(input)).isEqualTo(output);
    }
}
